+++ src/app/core/@ngrx/products/products.actions.ts
@@ -1,11 +1,11 @@
 import { createAction, props } from '@ngrx/store';
-import { IProduct } from 'src/app/feature/product/models';
+import { IProduct, ProductModel } from 'src/app/feature/product/models';
 
 export const getProducts = createAction('[Products] GET_PRODUCTS');
 
 export const getProductsSuccess = createAction(
     '[Products] GET_PRODUCTS_SUCCEESS',
-    props<{ products: IProduct[] }>()
+    props<{ products: ProductModel[] }>()
 );
 
 export const getProductsError = createAction(